+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - CircleCI badge in devctl's own README
 
+### Removed
+
+- Remove `gen kubeconfig` command
+
 ## [5.24.0] - 2023-05-02
 
 ### Changed